+++ src/BoclipsClient/__mocks__/AxiosBoclipsClient.ts
@@ -1,0 +1,9 @@
+import { BoclipsClient } from '../BoclipsClient';
+
+// noinspection JSUnusedGlobalSymbols
+export const AxiosBoclipsClient = jest.fn().mockImplementation(() => {
+  return {
+    retrieveVideo: jest.fn(),
+    createPlaybackEvent: jest.fn(),
+  } as BoclipsClient;
+});