+++ src/BoclipsPlayer/BoclipsPlayer.ts
@@ -1,10 +1,11 @@
 import deepmerge from 'deepmerge';
 import { addListener as addResizeListener } from 'resize-detector';
 import uuid from 'uuid/v1';
+import { AxiosBoclipsClient } from '../BoclipsClient/AxiosBoclipsClient';
+import { BoclipsClient } from '../BoclipsClient/BoclipsClient';
 import { ErrorHandler } from '../ErrorHandler/ErrorHandler';
 import { Analytics } from '../Events/Analytics';
 import { Video } from '../types/Video';
-import retrieveVideo from '../utils/retrieveVideo';
 import { Wrapper, WrapperConstructor } from '../Wrapper/Wrapper';
 import './BoclipsPlayer.less';
 import { BoclipsPlayerOptions, defaultOptions } from './BoclipsPlayerOptions';
@@ -22,8 +23,8 @@ export class BoclipsPlayer implements BoclipsPlayerInstance {
   private readonly container: HTMLElement;
   private readonly analytics: Analytics;
   private readonly errorHandler: ErrorHandler;
-  // @ts-ignore
-  private options: BoclipsPlayerOptions = {};
+  private readonly boclipsClient: BoclipsClient;
+  private options: Partial<BoclipsPlayerOptions> = {};
   private video: Video;
   private playerId: string = uuid();
 
@@ -58,8 +59,13 @@ export class BoclipsPlayer implements BoclipsPlayerInstance {
 
     container.classList.add('boclips-player-container');
 
-    this.analytics = new Analytics(this.playerId, this.options.analytics);
     this.errorHandler = new ErrorHandler(this.container);
+    this.boclipsClient = new AxiosBoclipsClient(this.options.boclips);
+    this.analytics = new Analytics(
+      this.boclipsClient,
+      this.playerId,
+      this.options.analytics,
+    );
 
     this.wrapper = new this.wrapperConstructor(
       container,
@@ -104,7 +110,8 @@ export class BoclipsPlayer implements BoclipsPlayerInstance {
 
     this.video = null;
 
-    return retrieveVideo(videoUri)
+    return this.boclipsClient
+      .retrieveVideo(videoUri)
       .then((video: Video) => {
         this.errorHandler.clearError();
 
@@ -113,6 +120,10 @@ export class BoclipsPlayer implements BoclipsPlayerInstance {
         this.wrapper.configureWithVideo(video);
       })
       .catch(error => {
+        if (this.errorHandler.isDefinedError(error)) {
+          this.errorHandler.handleError(error);
+          return;
+        }
         if (error && error.response && error.response.status) {
           if (error.response.status === 404) {
             this.errorHandler.handleError({
@@ -148,6 +159,8 @@ export class BoclipsPlayer implements BoclipsPlayerInstance {
   public getWrapper = () => this.wrapper;
 
   public getAnalytics = () => this.analytics;
+
+  public getBoclipsClient = () => this.boclipsClient;
 
   public getVideo = (): Video => this.video;