+++ src/ErrorHandler/__mocks__/ErrorHandler.ts
@@ -5,5 +5,6 @@ export const ErrorHandler = jest.fn().mockImplementation(() => {
   return {
     clearError: jest.fn(),
     handleError: jest.fn(),
+    isDefinedError: jest.fn().mockReturnValue(false),
   } as ErrorHandlerInstance;
 });